handle case when variable does not exist

+++ playmovie.sh
@@ -4,12 +4,17 @@
 
 movieFolder=$(getEnv.sh movieFolder)
 # player=${1-'mpv'} # pass video player as optional argument
-player=${1-'vlc --fullscreen'} # pass video player as optional argument
+player=${1-'vlc --fullscreen '} # pass video player as optional argument
 cd $movieFolder
 
 movie=$(find . -regex '.*.\(mp4\|mkv\|avi\)' |
   cut -d '/' -f2- |
   dmenu -i -l 20 -p "Choose a movie to play")
+
+if [ -z "$movie" ]; then
+  echo "No movie selected. Exiting.."
+  exit 0
+fi
 
 pgrep $player && killall $player
 $player "$movie" || notify-send "Error in playing $movie"
